# hello-world-service/src/main/resources/api-spec.yml
swagger: '2.0'
info:
  version: '0.1'
  title: 'Hello World API'
  description: 'The Hello World API .....'

schemes:
 - https
consumes:
 - application/json
produces:
 - application/json

parameters:
  user-name:
    name: user-name
    in: path
    description: The user name
    required: true
    type: string

paths:
  /greetings/{userName}:
    get:
      tags:
        - greetings
      summary: Get greeting
      description: Returns a greeting to the user.
      operationId: getHello
      parameters:
        - $ref: '#/parameters/user-name'
      responses:
        200:
          $ref: '#/responses/GreetingsResponse'
        400:
          $ref: '#/responses/BadRequest'
        500:
          $ref: '#/responses/ServerError'

#######################################################
# MODELS
#######################################################

definitions:
  GreetingWsDTO:
    description: 'The Greeting model'
    type: object
    properties:
      greeting:
        description: 'The greeting value.'
        type: string

  ErrorDetail:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: string
        description: "The error code, in screaming snake case: eg PAYLOAD_VALIDATION_ERROR"
        pattern: "^[A-Z]+(_[A-Z]+)*$"
      message:
        type: string
        description: "descriptive error message for debugging"
      moreInfo:
        type: string
        description: "link to documentation to investigate further and finding support"
    example:
      code: "INVALID_PARAMETER"
      message: "Missing required parameter"

  Error:
    type: object
    required:
      - status
      - code
    properties:
      status:
        type: integer
        description: "original HTTP error code, should be consistent with the response HTTP code"
        minimum: 100
        maximum: 599
      code:
        type: string
        description: "The error code, in screaming snake case: eg PAYLOAD_VALIDATION_ERROR"
        pattern: "^[A-Z]+(_[A-Z]+)*$"
      message:
        type: string
        description: "descriptive error message for debugging"
      moreInfo:
        type: string
        description: "link to documentation to investigate further and finding support"
      details:
        type: array
        items:
          $ref: '#/definitions/ErrorDetail'
#######################################################
# RESPONSES
#######################################################

responses:
  GreetingsResponse:
    description: The greetings response
    schema:
      $ref: '#/definitions/GreetingWsDTO'

  BadRequest:
    description: Indicates a bad request, which can be due to violation(s) of HTTP or API rules
    schema:
      $ref: '#/definitions/Error'
    examples:
      application/json:
        status: 400
        code: "BAD_REQUEST"
        message: "The locale is not recognized."

  ServerError:
    description: Indicates an unknown internal error that prevents the server from processing the request
    schema:
      $ref: '#/definitions/Error'
    examples:
      application/json:
        statusCode: 500
        type: "UNKNOWN_ERROR"
        message: "Internal server error."

host: virtserver.swaggerhub.com
basePath: /team-support/helloworld/0.1